+++ packages/ui/src/spinner/types.ts
@@ -1,0 +1,3 @@
+import { type ComponentProps } from '../utils/types';
+
+export type SpinnerProps = ComponentProps< 'svg' >;